Extract image conversion code out of dataset.py
dataset.py had grown very large, and Dataset.from_images()/
add_layer_from_images() plus their supporting helpers (ConversionLayerMapping,
_has_image_z_dimension, the ChunkedImages/PimsImages dispatch machinery
added over the last several commits) made up a large, mostly
self-contained chunk of it (~700 lines). Moved this into a new
webknossos/dataset/image_conversion.py module:

- ConversionLayerMapping: now a top-level Enum (was nested in Dataset).
  Dataset.ConversionLayerMapping is kept as a class attribute alias
  pointing at it, since it's public API (used externally, e.g. in
  tests as Dataset.ConversionLayerMapping.ENFORCE_SINGLE_LAYER).
- _has_image_z_dimension, _ImageConversionSetup,
  _prepare_chunked_image_conversion, _prepare_pims_image_conversion,
  SAFE_LARGE_XY: moved as-is.
- from_images()/add_layer_from_images(): moved as free functions taking
  the Dataset class/instance as an explicit first parameter
  (dataset_cls/dataset) instead of cls/self. Dataset.from_images and
  Dataset.add_layer_from_images become thin wrappers that keep their
  full signatures and docstrings (the public API/docs), just
  delegating to image_conversion.from_images(cls, ...) /
  image_conversion.add_layer_from_images(self, ...).

image_conversion.py uses `from __future__ import annotations` and a
TYPE_CHECKING-guarded import of Dataset to avoid a circular import
(dataset.py imports the module at runtime; image_conversion.py only
needs Dataset for type hints).

One non-obvious fix needed along the way: mypy rejected
ConversionLayerMapping used bare in from_images()'s signature once it
was only reachable via the Dataset.ConversionLayerMapping class
attribute assignment ("Variable ... is not valid as a type") - class
body attribute assignments aren't recognized as type aliases. Fixed by
referencing the fully-qualified image_conversion.ConversionLayerMapping
directly in the annotation and default value instead of relying on the
bare name resolving through the class body's own namespace.

dataset.py: 2701 -> 2032 lines.

Verified: full tests/dataset/test_add_layer_from_images.py +
test_from_images.py suite passes unchanged (34 passed, only the
pre-existing JVM/DICOM test fails, unrelated). mypy clean across all 47
files in webknossos/dataset/. Broader tests/dataset/ sweep: the only
failures are test_remote_dataset.py tests hitting the real
webknossos.org API (this scratch environment has no local WK test
server configured), none touching conversion code.
