+++ webknossos/webknossos/dataset/dataset.py
@@ -3,16 +3,13 @@ import json
 import logging
 import warnings
 from collections.abc import Callable, Iterable, Mapping, Sequence
-from enum import Enum, unique
-from itertools import product
 from os import PathLike
 from os.path import relpath
-from typing import TYPE_CHECKING, Any, NamedTuple, Union, cast
+from typing import TYPE_CHECKING, Any, Union, cast
 
 import attr
 import numpy as np
 from cluster_tools import Executor
-from natsort import natsort_keygen
 from numpy.typing import DTypeLike
 from upath import UPath
 
@@ -30,12 +27,7 @@ from ..geometry import (
 )
 from ..geometry.mag import MagLike
 from ..geometry.nd_bounding_box import derive_nd_bounding_box_from_shape
-from ._utils import pims_images
-from ._utils.chunked_images import (
-    ChunkedImages,
-    get_valid_chunked_image_suffixes,
-    try_open_chunked_images,
-)
+from . import image_conversion
 from .abstract_dataset import (
     DEFAULT_VERSION,
     AbstractDataset,
@@ -44,7 +36,6 @@ from .abstract_dataset import (
     _dtype_maybe,
 )
 from .defaults import (
-    DEFAULT_CHUNKS_PER_SHARD_FROM_IMAGES,
     DEFAULT_DATA_FORMAT,
     DEFAULT_DTYPE,
     PROPERTIES_FILE_NAME,
@@ -60,10 +51,8 @@ from .layer import (
     Zarr3Config,
 )
 from .layer.abstract_layer import (
-    _UNALLOWED_LAYER_NAME_CHARS,
     _validate_layer_name,
 )
-from .layer.layer import _get_shard_and_chunk_shapes
 from .ome_metadata import write_ome_metadata
 from .remote_dataset import RemoteAccessMode, RemoteDataset
 from .remote_folder import RemoteFolder
@@ -72,8 +61,6 @@ from .transfer_mode import TransferMode
 
 if TYPE_CHECKING:
     import pims
-
-    from .layer.view import MagView
 
 
 from webknossos.dataset.layer import (
@@ -101,18 +88,12 @@ from ..utils import (
     dump_path,
     enrich_path,
     is_fs_path,
-    named_partial,
     rmtree,
     strip_trailing_slash,
-    wait_and_ensure_success,
     warn_deprecated,
     wrap_executor,
 )
 from ._utils.infer_bounding_box_existing_files import infer_bounding_box_existing_files
-from ._utils.segmentation_recognition import (
-    guess_category_from_view,
-    guess_if_segmentation_path,
-)
 
 logger = logging.getLogger(__name__)
 
@@ -135,142 +116,6 @@ _ALLOWED_SEGMENTATION_LAYER_DTYPES = (
     "int32",
     "int64",
 )
-
-SAFE_LARGE_XY: int = 10_000_000_000  # 10 billion
-
-
-def _has_image_z_dimension(
-    filepath: UPath,
-    use_bioformats: bool,
-    is_segmentation: bool,
-) -> bool:
-    # Formats handled by a registered ChunkedImages subclass (e.g. .ims) are
-    # never probed through PimsImages/pims — they know their exact
-    # expected_bbox directly, and this is the only way such files are read.
-    chunked_images = try_open_chunked_images(
-        filepath,
-        channel=None,
-        timepoint=None,
-        swap_xy=False,
-        flip_x=False,
-        flip_y=False,
-        flip_z=False,
-        is_segmentation=is_segmentation,
-    )
-    if chunked_images is not None:
-        return chunked_images.expected_bbox.get_shape("z") > 1
-    return pims_images.has_image_z_dimension(
-        filepath, use_bioformats=use_bioformats, is_segmentation=is_segmentation
-    )
-
-
-class _ImageConversionSetup(NamedTuple):
-    """Everything add_layer_from_images needs to actually run the conversion,
-    computed differently for ChunkedImages vs. PimsImages sources by
-    _prepare_chunked_image_conversion / _prepare_pims_image_conversion. This
-    covers the parts of the two paths that differ (bounding box, mag_view,
-    per-chunk function, batch_size) — the surrounding layer/mag setup and
-    execution machinery stay shared in add_layer_from_images itself.
-    """
-
-    mag_view: "MagView"
-    func_per_chunk: Callable[..., tuple[tuple[int, int], int | None]]
-    batch_size: int | None
-
-
-def _prepare_chunked_image_conversion(
-    image_source: ChunkedImages,
-    layer: Layer,
-    *,
-    mag: Mag,
-    mag1_expected_bbox: NDBoundingBox,
-    chunk_shape: Vec3Int,
-    shard_shape: Vec3Int,
-    compress: bool,
-    current_dtype: np.dtype,
-    batch_size: int | None,
-) -> _ImageConversionSetup:
-    # ChunkedImages formats know their exact bounding box from metadata
-    # alone, so no placeholder inflation (and no post-hoc correction
-    # afterward) is needed.
-    layer.bounding_box = mag1_expected_bbox
-
-    mag_view = layer.add_mag(
-        mag=mag,
-        chunk_shape=chunk_shape,
-        shard_shape=shard_shape,
-        compress=compress,
-    )
-
-    func_per_chunk = named_partial(
-        image_source.read_chunk,
-        mag_view=mag_view,
-        dtype=current_dtype,
-    )
-
-    return _ImageConversionSetup(
-        mag_view=mag_view,
-        func_per_chunk=func_per_chunk,
-        batch_size=batch_size,  # unused for this path, passed through unchanged
-    )
-
-
-def _prepare_pims_image_conversion(
-    image_source: pims_images.PimsImages,
-    layer: Layer,
-    *,
-    mag: Mag,
-    mag1_expected_bbox: NDBoundingBox,
-    chunk_shape: Vec3Int,
-    shard_shape: Vec3Int,
-    compress: bool,
-    current_dtype: np.dtype,
-    batch_size: int | None,
-) -> _ImageConversionSetup:
-    # Setting a large enough bounding box, because the exact bounding box
-    # cannot be know a priori all the time. It will be replaced with the
-    # correct bounding box after reading through all actual images.
-    safe_size = mag1_expected_bbox.size.with_replaced(
-        mag1_expected_bbox.axes.index("x"), SAFE_LARGE_XY
-    ).with_replaced(mag1_expected_bbox.axes.index("y"), SAFE_LARGE_XY)
-    layer.bounding_box = mag1_expected_bbox.with_size(safe_size)
-
-    mag_view = layer.add_mag(
-        mag=mag,
-        chunk_shape=chunk_shape,
-        shard_shape=shard_shape,
-        compress=compress,
-    )
-
-    if batch_size is None:
-        if compress or (layer.data_format in (DataFormat.Zarr3, DataFormat.Zarr)):
-            # if data is compressed or dataformat is zarr, parallel write access
-            # to a shard leads to corrupted data, the batch size must be aligned
-            # with the shard size
-            batch_size = mag_view.info.shard_shape.z
-        else:
-            # in uncompressed wkw only writing to the same chunk is problematic
-            batch_size = mag_view.info.chunk_shape.z
-    elif compress or (layer.data_format in (DataFormat.Zarr3, DataFormat.Zarr)):
-        assert batch_size % mag_view.info.shard_shape.z == 0, (
-            f"batch_size {batch_size} must be divisible by z shard-size {mag_view.info.shard_shape.z} when creating compressed layers"
-        )
-    else:
-        assert batch_size % mag_view.info.chunk_shape.z == 0, (
-            f"batch_size {batch_size} must be divisible by z chunk-size {mag_view.info.chunk_shape.z}"
-        )
-
-    func_per_chunk = named_partial(
-        image_source.copy_to_view,
-        mag_view=mag_view,
-        dtype=current_dtype,
-    )
-
-    return _ImageConversionSetup(
-        mag_view=mag_view,
-        func_per_chunk=func_per_chunk,
-        batch_size=batch_size,
-    )
 
 
 def _find_array_info(layer_path: UPath) -> ArrayInfo | None:
@@ -313,116 +158,10 @@ class Dataset(AbstractDataset[Layer, SegmentationLayer]):
             ```
     """
 
-    @unique
-    class ConversionLayerMapping(Enum):
-        """Strategies for mapping file paths to layers when importing images.
-
-        These strategies determine how input image files are grouped into layers during
-        dataset creation using `Dataset.from_images()`. If no strategy is provided,
-        `INSPECT_SINGLE_FILE` is used as the default.
-
-        If none of the pre-defined strategies fit your needs, you can provide a custom
-        callable that takes a Path and returns a layer name string.
-
-        Examples:
-            Using default strategy:
-                ```
-                ds = Dataset.from_images("images/", "dataset/")
-                ```
-
-            Explicit strategy:
-                ```
-                ds = Dataset.from_images(
-                    "images/",
-                    "dataset/",
-                    map_filepath_to_layer_name=ConversionLayerMapping.ENFORCE_SINGLE_LAYER
-                )
-                ```
-
-            Custom mapping function:
-                ```
-                ds = Dataset.from_images(
-                    "images/",
-                    "dataset/",
-                    map_filepath_to_layer_name=lambda p: p.stem
-                )
-                ```
-        """
-
-        INSPECT_SINGLE_FILE = "inspect_single_file"
-        """Default strategy. Inspects first image file to determine if data is 2D or 3D.
-        For 2D data uses ENFORCE_LAYER_PER_FOLDER, for 3D uses ENFORCE_LAYER_PER_FILE."""
-
-        INSPECT_EVERY_FILE = "inspect_every_file"
-        """Like INSPECT_SINGLE_FILE but determines strategy separately for each file.
-        More flexible but slower for many files."""
-
-        ENFORCE_LAYER_PER_FILE = "enforce_layer_per_file"
-        """Creates a new layer for each input file. Useful for converting multiple
-        3D images or when each 2D image should become its own layer."""
-
-        ENFORCE_SINGLE_LAYER = "enforce_single_layer"
-        """Combines all input files into a single layer. Only useful when all
-        images are 2D slices that should be combined."""
-
-        ENFORCE_LAYER_PER_FOLDER = "enforce_layer_per_folder"
-        """Groups files by their containing folder. Each folder becomes one layer.
-        Useful for organized 2D image stacks."""
-
-        ENFORCE_LAYER_PER_TOPLEVEL_FOLDER = "enforce_layer_per_toplevel_folder"
-        """Groups files by their top-level folder. Useful when multiple layers each
-        have their stacks split across subfolders."""
-
-        def _to_callable(
-            self,
-            input_path: UPath,
-            input_files: Sequence[UPath],
-            use_bioformats: bool,
-        ) -> Callable[[UPath], str]:
-            ConversionLayerMapping = Dataset.ConversionLayerMapping
-
-            if self == ConversionLayerMapping.ENFORCE_LAYER_PER_FILE:
-                return lambda p: p.as_posix().replace("/", "_")
-            elif self == ConversionLayerMapping.ENFORCE_SINGLE_LAYER:
-                return lambda _p: input_path.name
-            elif self == ConversionLayerMapping.ENFORCE_LAYER_PER_FOLDER:
-                return lambda p: (
-                    input_path.name
-                    if p.parent == UPath()
-                    else p.parent.as_posix().replace("/", "_")
-                )
-            elif self == ConversionLayerMapping.ENFORCE_LAYER_PER_TOPLEVEL_FOLDER:
-                return lambda p: input_path.name if p.parent == UPath() else p.parts[0]
-            elif self == ConversionLayerMapping.INSPECT_EVERY_FILE:
-                # If a file has z dimensions, it becomes its own layer,
-                # if it's 2D, the folder becomes a layer.
-                return lambda p: (
-                    str(p)
-                    if _has_image_z_dimension(
-                        input_path / p,
-                        use_bioformats=use_bioformats,
-                        is_segmentation=guess_if_segmentation_path(p),
-                    )
-                    else (
-                        input_path.name
-                        if p.parent == UPath()
-                        else p.parent.as_posix().replace("/", "_")
-                    )
-                )
-            elif self == ConversionLayerMapping.INSPECT_SINGLE_FILE:
-                # As before, but only a single image is inspected to determine 2D vs 3D.
-                if _has_image_z_dimension(
-                    input_path / input_files[0],
-                    use_bioformats=use_bioformats,
-                    is_segmentation=guess_if_segmentation_path(input_files[0]),
-                ):
-                    return str
-                else:
-                    return lambda p: (
-                        input_path.name if p.parent == UPath() else p.parts[-2]
-                    )
-            else:
-                raise ValueError(f"Got unexpected ConversionLayerMapping value: {self}")
+    # Kept as a class attribute for backward compatibility: Dataset.ConversionLayerMapping
+    # used to be a nested class; the implementation now lives in image_conversion.py
+    # alongside from_images()/add_layer_from_images(), which use it directly.
+    ConversionLayerMapping = image_conversion.ConversionLayerMapping
 
     def __init__(
         self,
@@ -956,8 +695,10 @@ class Dataset(AbstractDataset[Layer, SegmentationLayer]):
         voxel_size: tuple[float, float, float] | None = None,
         name: str | None = None,
         *,
-        map_filepath_to_layer_name: ConversionLayerMapping
-        | Callable[[UPath], str] = ConversionLayerMapping.INSPECT_SINGLE_FILE,
+        map_filepath_to_layer_name: image_conversion.ConversionLayerMapping
+        | Callable[
+            [UPath], str
+        ] = image_conversion.ConversionLayerMapping.INSPECT_SINGLE_FILE,
         z_slices_sort_key: Callable[[UPath], Any] | None = None,
         voxel_size_with_unit: VoxelSize | None = None,
         layer_name: str | None = None,
@@ -1030,131 +771,31 @@ class Dataset(AbstractDataset[Layer, SegmentationLayer]):
             This method needs extra packages like tifffile or pylibczirw.
             Install with `pip install "webknossos[all]"`.
         """
-
-        input_upath = UPath(input_path)
-
-        valid_suffixes = pims_images.get_valid_pims_suffixes()
-        valid_suffixes.update(get_valid_chunked_image_suffixes())
-        if use_bioformats is not False:
-            valid_suffixes.update(pims_images.get_valid_bioformats_suffixes())
-
-        if z_slices_sort_key is None:
-            z_slices_sort_key = natsort_keygen()
-
-        if input_upath.is_file():
-            if input_upath.suffix.lstrip(".").lower() in valid_suffixes:
-                input_files = [UPath(input_upath.name)]
-                input_upath = input_upath.parent
-        else:
-            input_files = [
-                i.relative_to(input_upath)
-                for i in input_upath.glob("**/*")
-                if i.is_file() and i.suffix.lstrip(".").lower() in valid_suffixes
-            ]
-
-        if len(input_files) == 0:
-            raise ValueError(
-                "Could not find any supported image data. "
-                + f"The following suffixes are supported: {sorted(valid_suffixes)}"
-            )
-
-        if isinstance(map_filepath_to_layer_name, Dataset.ConversionLayerMapping):
-            with warnings.catch_warnings():
-                warnings.filterwarnings(
-                    "ignore",
-                    category=UserWarning,
-                    module="pims",
-                )
-                warnings.filterwarnings(
-                    "once",
-                    category=UserWarning,
-                    module="pims_images",
-                )
-                map_filepath_to_layer_name_func = (
-                    map_filepath_to_layer_name._to_callable(
-                        input_upath,
-                        input_files=input_files,
-                        use_bioformats=use_bioformats,
-                    )
-                )
-        else:
-            map_filepath_to_layer_name_func = map_filepath_to_layer_name
-        if voxel_size_with_unit is None:
-            assert voxel_size is not None, (
-                "Please supply either voxel_size or voxel_size_with_unit."
-            )
-            voxel_size_with_unit = VoxelSize(voxel_size)
-        else:
-            assert voxel_size is None, (
-                "Please supply either voxel_size or voxel_size_with_unit not both."
-            )
-
-        ds = cls(output_path, voxel_size_with_unit=voxel_size_with_unit, name=name)
-
-        filepaths_per_layer: dict[str, list[UPath]] = {}
-        for input_file in input_files:
-            layer_name_from_mapping = map_filepath_to_layer_name_func(input_file)
-            # Remove characters from layer name that are not allowed
-            layer_name_from_mapping = _UNALLOWED_LAYER_NAME_CHARS.sub(
-                "", layer_name_from_mapping
-            )
-            # Ensure layer name does not start with a dot
-            layer_name_from_mapping = layer_name_from_mapping.lstrip(".")
-
-            assert layer_name_from_mapping != "", (
-                f"Could not determine a layer name for {input_file}."
-            )
-
-            filepaths_per_layer.setdefault(layer_name_from_mapping, []).append(
-                input_upath / input_file
-            )
-
-        if layer_name is not None:
-            if len(filepaths_per_layer) == 1:
-                filepaths_per_layer[layer_name] = filepaths_per_layer.pop(
-                    layer_name_from_mapping
-                )
-            else:
-                filepaths_per_layer = {
-                    f"{layer_name}_{k}": v for k, v in filepaths_per_layer.items()
-                }
-        with wrap_executor(executor) as executor:
-            with warnings.catch_warnings():
-                warnings.filterwarnings(
-                    "ignore",
-                    category=UserWarning,
-                    module="pims_images",
-                )
-                warnings.filterwarnings(
-                    "ignore",
-                    category=UserWarning,
-                    module="pims",
-                )
-                for layer_name, filepaths in filepaths_per_layer.items():
-                    filepaths.sort(key=z_slices_sort_key)
-
-                    ds.add_layer_from_images(
-                        filepaths[0] if len(filepaths) == 1 else filepaths,
-                        layer_name,
-                        category=layer_category,
-                        data_format=data_format,
-                        chunk_shape=chunk_shape,
-                        shard_shape=shard_shape,
-                        chunks_per_shard=chunks_per_shard,
-                        compress=compress,
-                        swap_xy=swap_xy,
-                        flip_x=flip_x,
-                        flip_y=flip_y,
-                        flip_z=flip_z,
-                        use_bioformats=use_bioformats,
-                        batch_size=batch_size,
-                        allow_multiple_layers=True,
-                        max_layers=max_layers - len(ds.layers),
-                        truncate_rgba_to_rgb=False,
-                        executor=executor,
-                    )
-
-        return ds
+        return image_conversion.from_images(
+            cls,
+            input_path,
+            output_path,
+            voxel_size,
+            name,
+            map_filepath_to_layer_name=map_filepath_to_layer_name,
+            z_slices_sort_key=z_slices_sort_key,
+            voxel_size_with_unit=voxel_size_with_unit,
+            layer_name=layer_name,
+            layer_category=layer_category,
+            data_format=data_format,
+            chunk_shape=chunk_shape,
+            shard_shape=shard_shape,
+            chunks_per_shard=chunks_per_shard,
+            compress=compress,
+            swap_xy=swap_xy,
+            flip_x=flip_x,
+            flip_y=flip_y,
+            flip_z=flip_z,
+            use_bioformats=use_bioformats,
+            max_layers=max_layers,
+            batch_size=batch_size,
+            executor=executor,
+        )
 
     def add_layer(
         self,
@@ -1520,343 +1161,33 @@ class Dataset(AbstractDataset[Layer, SegmentationLayer]):
         * `executor`: pass a `ClusterExecutor` instance to parallelize the conversion jobs across the batches
         """
 
-        _validate_layer_name(layer_name)
-        if category is None:
-            image_path_for_category_guess: UPath
-            if (
-                isinstance(images, str)
-                or isinstance(images, PathLike)
-                or isinstance(images, UPath)
-            ):
-                image_path_for_category_guess = UPath(images)
-            else:
-                image_path_for_category_guess = UPath(images[0])
-            category = (
-                "segmentation"
-                if guess_if_segmentation_path(image_path_for_category_guess)
-                else "color"
-            )
-            user_set_category = False
-        else:
-            user_set_category = True
-
-        def _open_chunked_images(open_kwargs: dict) -> ChunkedImages | None:
-            # Chunk-based formats (currently only .ims) know their exact
-            # bounding box from metadata alone and read/write whole
-            # shard-sized blocks directly. This is the only way such formats
-            # are read — use_bioformats has no effect on them.
-            return try_open_chunked_images(
-                images,
-                channel=open_kwargs.get("channel", channel),
-                timepoint=open_kwargs.get("timepoint", timepoint),
-                swap_xy=swap_xy,
-                flip_x=flip_x,
-                flip_y=flip_y,
-                flip_z=flip_z,
-                is_segmentation=category == "segmentation",
-            )
-
-        def _reopen_image_source(
-            open_kwargs: dict,
-        ) -> "pims_images.PimsImages | ChunkedImages":
-            chunked = _open_chunked_images(open_kwargs)
-            if chunked is not None:
-                return chunked
-            return pims_images.PimsImages(
-                images,
-                swap_xy=swap_xy,
-                flip_x=flip_x,
-                flip_y=flip_y,
-                flip_z=flip_z,
-                use_bioformats=use_bioformats,
-                is_segmentation=category == "segmentation",
-                **open_kwargs,
-            )
-
-        image_source: pims_images.PimsImages | ChunkedImages
-        chunked_image_source = _open_chunked_images(
-            {"channel": channel, "timepoint": timepoint}
+        return image_conversion.add_layer_from_images(
+            self,
+            images,
+            layer_name,
+            category,
+            data_format=data_format,
+            mag=mag,
+            chunk_shape=chunk_shape,
+            shard_shape=shard_shape,
+            chunks_per_shard=chunks_per_shard,
+            compress=compress,
+            topleft=topleft,
+            swap_xy=swap_xy,
+            flip_x=flip_x,
+            flip_y=flip_y,
+            flip_z=flip_z,
+            dtype=dtype,
+            use_bioformats=use_bioformats,
+            channel=channel,
+            timepoint=timepoint,
+            czi_channel=czi_channel,
+            batch_size=batch_size,
+            allow_multiple_layers=allow_multiple_layers,
+            max_layers=max_layers,
+            truncate_rgba_to_rgb=truncate_rgba_to_rgb,
+            executor=executor,
         )
-        if chunked_image_source is not None:
-            image_source = chunked_image_source
-        else:
-            image_source = pims_images.PimsImages(
-                images,
-                channel=channel,
-                timepoint=timepoint,
-                czi_channel=czi_channel,
-                swap_xy=swap_xy,
-                flip_x=flip_x,
-                flip_y=flip_y,
-                flip_z=flip_z,
-                use_bioformats=use_bioformats,
-                is_segmentation=category == "segmentation",
-            )
-        possible_layers = image_source.get_possible_layers()
-        # Check if 4 color channels should be converted to
-        # 3 color channels (rbg)
-        if (
-            possible_layers is not None
-            and truncate_rgba_to_rgb
-            and len(possible_layers.get("channel", [])) == 4
-        ):
-            # Remove channels from possible_layers to keep those
-            # and automatically truncate to 3 channels
-            # (pims_images takes care of this:)
-            del possible_layers["channel"]
-        # Further below, we iterate over suffix_with_pims_open_kwargs_per_layer in the for-loop
-        # to add one layer per possible_layer if allow_multiple_layers is True.
-        # If just a single layer is added, we still add a default value in the dict.
-        if possible_layers is not None and len(possible_layers) > 0:
-            if allow_multiple_layers:
-                # Get all combinations of possible layers. E.g.
-                # possible_layers = {
-                #    "channel": [0, 1, 3, 4, 5],
-                #    "timepoint": [0, 1],
-                # }
-                # suffix_with_pims_open_kwargs_per_layer = {
-                #    "__channel0_timepoint0", {"channel": 0, "timepoint": 0},
-                #    "__channel0_timepoint1", {"channel": 0, "timepoint": 1},
-                #    "__channel0_timepoint2", {"channel": 0, "timepoint": 2},
-                #    …,
-                #    "__channel1_timepoint0", {"channel": 1, "timepoint": 0},
-                #    …,
-                # }
-                suffix_with_pims_open_kwargs_per_layer = {
-                    "__" + "_".join(f"{k}{v}" for k, v in sorted(pairs)): dict(pairs)
-                    for pairs in product(
-                        *(
-                            [(key, value) for value in values]
-                            for key, values in possible_layers.items()
-                        )
-                    )
-                }
-            else:
-                # initialize PimsImages as above, with normal layer name
-                suffix_with_pims_open_kwargs_per_layer = {"": {}}
-                warnings.warn(
-                    f"[INFO] There are dimensions beyond channels and xyz which cannot be read: {possible_layers}. "
-                    "Defaulting to the first one. "
-                    "Please set allow_multiple_layers=True if all of them should be written to different layers, "
-                    "or set specific values as arguments.",
-                )
-        else:
-            # initialize PimsImages as above, with normal layer name
-            suffix_with_pims_open_kwargs_per_layer = {"": {}}
-        first_layer = None
-        add_layer_kwargs = {}
-        if category == "segmentation":
-            add_layer_kwargs["largest_segment_id"] = 0
-        if len(suffix_with_pims_open_kwargs_per_layer) > max_layers:
-            warnings.warn(
-                f"[INFO] Limiting the number of added layers to {max_layers} out of {len(suffix_with_pims_open_kwargs_per_layer)}. "
-                + "Please increase `max_layers` if you want more layers to be added.",
-            )
-        for _, (
-            layer_name_suffix,
-            pims_open_kwargs,
-        ) in zip(range(max_layers), suffix_with_pims_open_kwargs_per_layer.items()):
-            # If pims_open_kwargs is empty there's no need to re-open the images:
-            if len(pims_open_kwargs) > 0:
-                # Set parameters from this method as default
-                # if they are not part of the kwargs per layer:
-                pims_open_kwargs.setdefault("timepoint", timepoint)  # type: ignore
-                pims_open_kwargs.setdefault("channel", channel)  # type: ignore
-                pims_open_kwargs.setdefault("czi_channel", czi_channel)  # type: ignore
-                image_source = _reopen_image_source(pims_open_kwargs)
-            if dtype is None:
-                current_dtype = np.dtype(image_source.dtype)
-                if current_dtype.byteorder == ">":
-                    current_dtype = current_dtype.newbyteorder("<")
-            else:
-                current_dtype = np.dtype(dtype)
-
-            layer = self.add_layer(
-                layer_name=layer_name + layer_name_suffix,
-                category=category,
-                data_format=data_format,
-                dtype=current_dtype,
-                num_channels=image_source.num_channels,
-                **add_layer_kwargs,  # type: ignore[arg-type]
-            )
-
-            expected_bbox = image_source.expected_bbox
-
-            _shard_shape_user_specified = (
-                shard_shape is not None or chunks_per_shard is not None
-            )
-            chunk_shape, shard_shape = _get_shard_and_chunk_shapes(
-                data_format=layer.data_format,
-                layer_bounding_box=None,
-                chunk_shape=chunk_shape,
-                chunks_per_shard=chunks_per_shard,
-                shard_shape=shard_shape,
-            )
-            # For Zarr3 image imports, default to z-aligned shards for efficient
-            # slice-by-slice writing, regardless of total z size.
-            if (
-                not _shard_shape_user_specified
-                and layer.data_format == DataFormat.Zarr3
-            ):
-                shard_shape = chunk_shape * DEFAULT_CHUNKS_PER_SHARD_FROM_IMAGES
-            # When the expected bbox is 2D the chunk_shape is set to 2D too.
-            if (
-                expected_bbox.get_shape("z") == 1
-                and layer.data_format == DataFormat.Zarr3
-            ):
-                chunk_shape = chunk_shape.with_z(1)
-                shard_shape = shard_shape.with_z(1)
-
-            mag = Mag(mag)
-            mag1_expected_bbox = expected_bbox.from_mag_to_mag1(mag).offset(topleft)
-
-            prepare_conversion = (
-                _prepare_chunked_image_conversion
-                if isinstance(image_source, ChunkedImages)
-                else _prepare_pims_image_conversion
-            )
-            mag_view, func_per_chunk, batch_size = prepare_conversion(
-                image_source,  # type: ignore[arg-type]
-                layer,
-                mag=mag,
-                mag1_expected_bbox=mag1_expected_bbox,
-                chunk_shape=chunk_shape,
-                shard_shape=shard_shape,
-                compress=compress,
-                current_dtype=current_dtype,
-                batch_size=batch_size,
-            )
-
-            if (
-                additional_axes := set(layer.bounding_box.axes).difference(
-                    "c", "x", "y", "z"
-                )
-            ) and layer.data_format == DataFormat.WKW:
-                if all(
-                    layer.bounding_box.get_shape(axis) == 1 for axis in additional_axes
-                ):
-                    warnings.warn(
-                        f"[INFO] The data has additional axes {additional_axes}, but they are all of size 1. "
-                        + "These axes are not stored in the layer."
-                    )
-                    layer.bounding_box = BoundingBox.from_ndbbox(layer.bounding_box)
-                else:
-                    raise RuntimeError(
-                        f"WKW datasets only support x, y, z axes, got {additional_axes}. Please use `data_format='zarr3'` instead."
-                    )
-
-            if isinstance(image_source, ChunkedImages):
-                # Split into full 3D shard-aligned chunks so each job reads
-                # exactly one shard's worth of data and writes it directly.
-                # layer.bounding_box is already the exact bbox (no placeholder
-                # inflation) for ChunkedImages, so chunking it directly is safe.
-                chunked_shard_shape = mag_view.info.shard_shape
-                args = list(
-                    layer.bounding_box.chunk(chunked_shard_shape, chunked_shard_shape)
-                )
-            else:
-                # _prepare_pims_image_conversion always resolves batch_size to
-                # a concrete int; only the (unused-here) ChunkedImages path
-                # leaves it as None.
-                assert batch_size is not None
-                buffered_slice_writer_shape = layer.bounding_box.size_xyz.with_z(
-                    batch_size
-                )
-                args = list(
-                    layer.bounding_box.chunk(
-                        buffered_slice_writer_shape,
-                        Vec3Int(1, 1, batch_size),
-                    )
-                )
-
-            with warnings.catch_warnings():
-                # We need to catch and ignore a warning here about comparing persisted properties.
-                # The problem is that there is an `axisOrder` property that goes in the datasource-properties.json
-                # file, but is only stored in a mag object. However, at first we don't have any mags
-                # so there is no place to store them. When we add the first mag and update the properties,
-                # there is a check that reads the properties first and compares them to the current state.
-                # This check fails because the axisOrder property hasn't been stored in the properties file.
-                # It is safe to ignore this warning because it is only an intermediate problem in this function.
-                # At the end of this function, the properties are complete and consistent.
-                warnings.filterwarnings(
-                    "ignore",
-                    message=".* properties changed in a way that they are not comparable anymore. Most likely the bounding box naming or axis order changed.*",
-                    category=UserWarning,
-                    module="webknossos",
-                )
-                with wrap_executor(executor) as executor:
-                    shapes_and_max_ids = wait_and_ensure_success(
-                        executor.map_to_futures(func_per_chunk, args),
-                        executor=executor,
-                        progress_desc=f"Creating layer [bold blue]{layer.name}[/bold blue] from images",
-                    )
-                shapes, max_ids = zip(*shapes_and_max_ids)
-                if category == "segmentation":
-                    max_id = max(max_ids)
-                    cast(SegmentationLayer, layer).largest_segment_id = max_id
-                if not isinstance(image_source, ChunkedImages):
-                    # ChunkedImages formats already have the exact bbox set
-                    # above; correcting it from per-chunk shapes here would be
-                    # wrong for them anyway, since each job's shape reflects
-                    # only its own shard-sized chunk, not the full extent.
-                    layer.bounding_box = layer.bounding_box.with_size_xyz(
-                        Vec3Int(
-                            pims_images.dimwise_max(shapes)
-                            + (layer.bounding_box.get_shape("z"),)
-                        )
-                        * mag.to_vec3_int().with_z(1)
-                    )
-            if expected_bbox != layer.bounding_box:
-                warnings.warn(
-                    "[WARNING] Some images are larger than expected, smaller slices are padded with zeros now. "
-                    + f"New bbox is {layer.bounding_box}, expected {expected_bbox}."
-                )
-
-            # Check if category of layer is set correctly
-            try:
-                if not user_set_category:
-                    # When the category is not set by the user, we use a very simple heuristic to guess the category
-                    # based on the file name of the input images. After loading the images, we check if the guessed
-                    # category might be wrong and adjust it if necessary. This second heuristic is based on the
-                    # pixel data of the images
-                    guessed_category = guess_category_from_view(layer.get_finest_mag())
-                    if guessed_category != layer.category:
-                        new_layer_properties: LayerProperties
-                        if guessed_category == SEGMENTATION_CATEGORY:
-                            logger.info("The layer category is set to segmentation.")
-                            new_layer_properties = SegmentationLayerProperties(
-                                **(
-                                    attr.asdict(layer._properties, recurse=False)
-                                ),  # use all attributes from LayerProperties
-                                largest_segment_id=int(max(max_ids)),
-                            )
-                            new_layer_properties.category = SEGMENTATION_CATEGORY
-                            self._layers[layer.name] = SegmentationLayer(
-                                self, new_layer_properties, read_only=False
-                            )
-                        else:
-                            logger.info("The layer category is set to color.")
-                            _properties = attr.asdict(layer._properties, recurse=False)
-                            _properties.pop("largest_segment_id", None)
-                            _properties.pop("mappings", None)
-
-                            new_layer_properties = LayerProperties(**_properties)
-                            new_layer_properties.category = COLOR_CATEGORY
-                            self._layers[layer.name] = Layer(
-                                self, new_layer_properties, read_only=False
-                            )
-                        self._properties.update_for_layer(
-                            layer.name, new_layer_properties
-                        )
-                        self._save_dataset_properties()
-
-            except Exception:
-                # The used heuristic was not able to guess the layer category, the previous value is kept
-                pass
-            if first_layer is None:
-                first_layer = layer
-        assert first_layer is not None
-        return first_layer
 
     def write_layer(
         self,
